refactor: extract clear module
